# Conflicts:
#	server/src/main/java/_BE_Project/Score/ScoreRepository.java
#	server/src/main/java/_BE_Project/Score/ScoreService.java
#	server/src/main/java/_BE_Project/answer/AnswerController.java
#	server/src/main/java/_BE_Project/answer/AnswerMapper.java
#	server/src/main/java/_BE_Project/answer/dto/AnswerDto.java

+++ server/src/main/java/_BE_Project/question/QuestionService.java
@@ -1,53 +1,43 @@
 package _BE_Project.question;
 
-import _BE_Project.Score.ScoreService;
 import _BE_Project.exception.BusinessLogicException;
 import _BE_Project.exception.ExceptionCode;
 import _BE_Project.member.entity.Member;
-import _BE_Project.member.repository.MemberRepository;
 import _BE_Project.member.service.MemberService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Optional;
 
 @Transactional
 @Service
+@RequiredArgsConstructor
 public class QuestionService {
 
     private final QuestionRepository questionRepository;
     private final MemberService memberService;
 
-    private final ScoreService scoreService;
-
-    public QuestionService(QuestionRepository questionRepository,
-                           MemberService memberService, ScoreService scoreService) {
-        this.questionRepository = questionRepository;
-        this.memberService = memberService;
-        this.scoreService = scoreService;
-    }
-
     public Question createQuestion (Question question) {
-        memberService.findMember(question.getMember().getMemberId());
-        Question saveQuestion = questionRepository.save(question);
-        question.setScore(0);
-
-        return saveQuestion;
+        question.setMember(memberService.findByEmail());
+        return questionRepository.save(question);
     }
-
+    @Transactional
     public Question updateQuestion (Question question) {
-        memberService.findMember(question.getMember().getMemberId());
+
+        Member member = memberService.findByEmail();
         Question findQuestion = findVerifyQuestion(question.getQuestionId());
 
-        Optional.ofNullable(question.getTitle()).ifPresent(title -> findQuestion.setTitle(title));
-        Optional.ofNullable(question.getContent()).ifPresent(content -> findQuestion.setContent(content));
+        if(findQuestion.getMember().getMemberId() != member.getMemberId()){
+            throw new BusinessLogicException(ExceptionCode.QUESTION_UPDATE_NO_PERMISSION);
+        }
 
-        return questionRepository.save(findQuestion);
+        findQuestion.setTitle(question.getTitle());
+        findQuestion.setContent(question.getContent());
+        return findQuestion;
     }
 
     public Question findQuestion (long questionId) {
@@ -66,15 +56,24 @@ public class QuestionService {
 
     @Transactional
     public Page<Question> searchQuestion (String keyword, int page, int size) {
-        Page<Question> questions = questionRepository.findByTitleContaining(keyword, PageRequest.of(page, size, Sort.by("questionId").descending()));
+        Page<Question> questions = questionRepository.findByTitleContainingIgnoreCase(keyword, PageRequest.of(page, size, Sort.by("questionId").descending()));
         return questions;
     }
 
     public void deleteQuestion (long questionId) {
+
+        Member member = memberService.findByEmail();
+        Question findQuestion = findVerifyQuestion(questionId);
+
+        if(findQuestion.getMember().getMemberId() != member.getMemberId()){
+            throw new BusinessLogicException(ExceptionCode.QUESTION_DELETE_NO_PERMISSION);
+        }
+
         questionRepository.deleteById(questionId);
     }
 
-    private Question findVerifyQuestion (long questionId) {
+
+    public Question findVerifyQuestion (long questionId) {
         Optional<Question> optionalQuestion = questionRepository.findById(questionId);
         Question findQuestion = optionalQuestion.orElseThrow(()-> new BusinessLogicException(ExceptionCode.QUESTION_NOT_FOUND));
         return findQuestion;
